Adding function file and import to script 05
