+++ scripts/05_fit_heart_disease_model.py
@@ -20,42 +20,10 @@ from sklearn.linear_model import LogisticRegression
 from sklearn.metrics import ConfusionMatrixDisplay, classification_report
 from sklearn.exceptions import UndefinedMetricWarning
 warnings.filterwarnings("ignore", category=UndefinedMetricWarning)
+import sys
+sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+from src.mean_std_cross_val_scores import mean_std_cross_val_scores
 
-# --- Start of code block copied from another author ---
-# Title: Function to consolidate cross validation scores into a pandas series.
-# Author: Varada Kolhatkar & Michael Gelbart
-# Source: https://pages.github.ubc.ca/mds-2025-26/DSCI_571_sup-learn-1_students/README.html 
-# Taken from: DSCI-571: Laboratory 2
-def mean_std_cross_val_scores(model, X_train, y_train, **kwargs):
-    """
-    Returns mean and std of cross validation
-
-    Parameters
-    ----------
-    model :
-        scikit-learn model
-    X_train : numpy array or pandas DataFrame
-        X in the training data
-    y_train :
-        y in the training data
-
-    Returns
-    ----------
-        pandas Series with mean scores from cross_validation
-    """
-
-    scores = cross_validate(model, X_train, y_train, **kwargs)
-
-    mean_scores = pd.DataFrame(scores).mean()
-    std_scores = pd.DataFrame(scores).std()
-    out_col = []
-
-    for i in range(len(mean_scores)):
-        out_col.append((f"%0.3f (+/- %0.3f)" % (mean_scores.iloc[i], 
-                                                std_scores.iloc[i])))
-
-    return pd.Series(data=out_col, index=mean_scores.index)
-# --- End of code block copied from another author ---
 
 def build_pipe(preprocessor, model):
     """
